requirements.txt
absl-py==0.11.0
annotated-types==0.7.0
anyio==4.5.2
appnope==0.1.1
astroid==2.4.2
astunparse==1.6.3
Authlib==1.2.1
backcall==0.2.0
beautifulsoup4==4.10.0
blinker==1.8.2
bs4==0.0.1
cachetools==4.1.1
certifi==2020.11.8
cffi==1.16.0
chardet==3.0.4
click==8.1.7
colorama==0.4.4
configparser==5.0.2
coverage==7.6.1
crayons==0.4.0
cryptography==41.0.4
cycler==0.10.0
decorator==4.4.2
distro==1.9.0
dnspython==2.4.2
email-validator==2.0.0.post2
exceptiongroup==1.2.2
fake-useragent==1.5.1
Flask==2.2.5
Flask-Cors==3.0.10
Flask-JWT-Extended==4.6.0
flask-mongoengine==1.0.0
Flask-Script==2.0.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
gast==0.3.3
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.33.2
h11==0.14.0
h5py==3.10.0
httpcore==1.0.7
httpx==0.28.1
idna==2.10
importlib-metadata==6.8.0
importlib-resources==6.1.0
iniconfig==2.0.0
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
isort==5.6.4
itsdangerous==2.1.2
jedi==0.17.2
jieba==0.42.1
Jinja2==3.1.2
jiter==0.8.2
jupyter-client==6.1.7
jupyter-core==4.7.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
Markdown==3.3.3
MarkupSafe==2.1.3
matplotlib==3.3.4
mccabe==0.6.1
mongoengine==0.27.0
mongomock==4.3.0
numpy==1.24.1
oauthlib==3.1.0
openai==1.64.0
opt-einsum==3.3.0
packaging==23.2
pandas==1.1.4
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
pluggy==1.5.0
prompt-toolkit==3.0.8
protobuf==3.14.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.7.2
PyJWT==2.9.0
pylint==2.6.0
pymongo==4.5.0
pyparsing==2.4.7
PyPDF2==3.0.1
pytest==8.3.4
pytest-cov==5.0.0
pytest-mock==3.14.0
python-dateutil==2.8.1
python-dotenv==1.0.0
pytz==2020.4
PyYAML==5.3.1
pyzmq==20.0.0
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.5.4
selenium==3.141.0
sentinels==1.0.0
six==1.15.0
sniffio==1.3.1
soupsieve==2.2.1
SQLAlchemy==2.0.38
termcolor==1.1.0
toml==0.10.2
tomli==2.2.1
tornado==6.1
tqdm==4.67.1
traitlets==5.0.5
typing_extensions==4.12.2
urllib3==1.26.2
wcwidth==0.2.5
webdriver-manager==4.0.1
Werkzeug==3.0.6
wrapt==1.12.1
WTForms==3.1.0
z3-solver==4.8.12.0
zipp==3.17.0
